move start_demuxer_task back into demux.rs add file_extension to FileSinkConfig
