+++ datafusion/core/src/datasource/file_format/json.rs
@@ -30,7 +30,9 @@ use super::{
     DEFAULT_SCHEMA_INFER_MAX_RECORD,
 };
 use crate::datasource::file_format::file_compression_type::FileCompressionType;
-use crate::datasource::file_format::write::demux::DemuxedStreamReceiver;
+use crate::datasource::file_format::write::demux::{
+    start_demuxer_task, DemuxedStreamReceiver,
+};
 use crate::datasource::file_format::write::BatchSerializer;
 use crate::datasource::physical_plan::{FileGroupDisplay, FileSink};
 use crate::datasource::physical_plan::{FileSinkConfig, NdJsonExec};
@@ -390,7 +392,7 @@ impl DataSink for JsonSink {
     ) -> Result<u64> {
         let object_store = self.config.get_object_store(context)?;
         let (demux_task, file_stream_rx) =
-            self.config.start_demuxer_task(data, context, "json".into());
+            start_demuxer_task(&self.config, data, context);
         self.spawn_writer_tasks_and_join(
             context,
             demux_task,
