+++ datafusion/proto/src/physical_plan/from_proto.rs
@@ -656,6 +656,7 @@ impl TryFrom<&protobuf::FileSinkConfig> for FileSinkConfig {
             table_partition_cols,
             insert_op,
             keep_partition_by_columns: conf.keep_partition_by_columns,
+            file_extension: conf.file_extension.clone(),
         })
     }
 }
