+++ datafusion/proto/tests/cases/roundtrip_physical_plan.rs
@@ -1275,6 +1275,7 @@ fn roundtrip_json_sink() -> Result<()> {
         table_partition_cols: vec![("plan_type".to_string(), DataType::Utf8)],
         insert_op: InsertOp::Overwrite,
         keep_partition_by_columns: true,
+        file_extension: "json".into(),
     };
     let data_sink = Arc::new(JsonSink::new(
         file_sink_config,
@@ -1311,6 +1312,7 @@ fn roundtrip_csv_sink() -> Result<()> {
         table_partition_cols: vec![("plan_type".to_string(), DataType::Utf8)],
         insert_op: InsertOp::Overwrite,
         keep_partition_by_columns: true,
+        file_extension: "csv".into(),
     };
     let data_sink = Arc::new(CsvSink::new(
         file_sink_config,
@@ -1370,6 +1372,7 @@ fn roundtrip_parquet_sink() -> Result<()> {
         table_partition_cols: vec![("plan_type".to_string(), DataType::Utf8)],
         insert_op: InsertOp::Overwrite,
         keep_partition_by_columns: true,
+        file_extension: "parquet".into(),
     };
     let data_sink = Arc::new(ParquetSink::new(
         file_sink_config,
